src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Deprecated==1.2.13
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycparser==2.21
pydantic==1.10.5
PyGithub==1.58.0
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.15.0
yarl==1.8.2
